refactor: use implementation in user service
